+++ src/zip.h
@@ -12,7 +12,7 @@ namespace zip {
 bool Zip(const std::string& src_dir, const std::string& dest_file, std::string* error);
 
 //zip and encrypt from a buffer
-bool Zip(void* buf, const unsigned long buf_size, const std::string& relative_path, const std::string& dest_file, const char* password, std::string* error);
+bool Zip(void* buf, const size_t buf_size, const std::string& relative_path, const std::string& dest_file, const char* password, std::string* error);
 
 bool Unzip(const std::string& zip_file, const std::string& dest_file, std::string* error);
